#2 added examples package in genos
# added more advanced examples and restructured readme
