+++ src/genos/__init__.py
@@ -1,1 +1,3 @@
+from genos.instantiate import get_class  # noqa: F401
+from genos.instantiate import get_method  # noqa: F401
 from genos.instantiate import recursive_instantiate  # noqa: F401
